CMFIVE-1898 - AttachmentBoxWidget - исправлена ошибка в удалении вложений

+++ CM5/Core/gui-server-impl/src/main/java/ru/intertrust/cm/core/gui/impl/server/widget/AttachmentBoxHandler.java
@@ -150,10 +150,12 @@ public class AttachmentBoxHandler extends LinkEditingWidgetHandler {
 
         newObjects = processAttachmentItems(newlyAddedAttachments, newObjects, context, false);
 
-        List<AttachmentItem> deletedAttachments = attachmentBoxState.getNewlyDeletedAttachments();
-        for (AttachmentItem deletedAttachment : deletedAttachments) {
-            if (deletedAttachment.getId() != null) {
-                attachmentService.deleteAttachment(deletedAttachment.getId());
+        if (isSelectionMode(new FieldPath(context.getWidgetConfig().getFieldPathConfig().getValue()))) {
+            List<AttachmentItem> deletedAttachments = attachmentBoxState.getNewlyDeletedAttachments();
+            for (AttachmentItem deletedAttachment : deletedAttachments) {
+                if (deletedAttachment.getId() != null) {
+                    attachmentService.deleteAttachment(deletedAttachment.getId());
+                }
             }
         }
         return newObjects;
